♻️ refacor: 폴더 구조 변경 및 ui 수정
Resolves:#18
