+++ src/constants/Modals.ts
@@ -1,9 +1,7 @@
-import {
-  InputContent,
-  OnlyTextContent,
-  WarningContent,
-} from '@/components/common/Modals/contents';
-import type { ModalProps,ModalVariant } from '@/types/Modals';
+import { InputContent } from '@/components/common/Modals/contents/InputContent';
+import { OnlyTextContent } from '@/components/common/Modals/contents/OnlyTextContent';
+import { WarningContent } from '@/components/common/Modals/contents/WarningContent';
+import type { ModalProps, ModalVariant } from '@/types/Modals';
 
 type ContentMapType = {
   [V in ModalVariant]: React.FC<Extract<ModalProps, { variant: V }>>;
